Split client, add abc and class inheritance
